% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myr6.r
\name{MyR6Class}
\alias{MyR6Class}
\title{Create MyR6Class objects}
\arguments{
\item{a}{A number}

\item{b}{Another number}
}
\value{
An object of type \code{MyR6Class}
}
\description{
\code{MyR6Class$new} constructs the object and returns it
}
\details{
A longer description goes here.
}
\section{Usage}{

MyR6Class$new(a, b)
}

\examples{
x <- MyR6Class$new(1,2)

}
